+++ History.md
@@ -2,13 +2,14 @@
 ==================
   * Fix instrumentation for ES6 class members (https://github.com/tntim96/JSCover/issues/261)
   * Fix instrumentation for array definition across lines (https://github.com/tntim96/JSCover/issues/260)
+  * Internal: Upgrade Jasmine 2.6.4 to 2.7.0
 
 2.0.1 / 2017-07-23
 ==================
   * Convert JSON merger from Mozilla Rhino to Google GSON (https://github.com/tntim96/JSCover/issues/258)
 
 2.0.0 / 2017-07-16
-==================
+=======ends===========
   * Replace Rhino AST engine with Closure Compiler (https://github.com/tntim96/JSCover/issues/255)
   * Internal: Upgrade HtmlUnit 2.24 to 2.27, Mockito 2.7.5 to 2.8.47, Jasmine 2.5.2 to 2.6.4